Add support for subcmd config option
It is possible to use kup in conjunction with another authorization
system that already relies on ssh for authentication (e.g. gitolite),
in which case we need to be able to specify a subcommand to specifically
invoke the kup server.

If no subcmd is specified, the default standalone kup behaviour is used.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>